feat: support script for SURE-5880
This includes a support script that can be used to upgrade an EKS
cluster using the Rancher API. Its designed for Rancher 2.6.10 where
there is a UI bug that prevents users from doing this. This is a
temporary solution until the bug fix is released in 2.6.11.

See the README.md for usage instructions, requirements and a demo.

Signed-off-by: Richard Case <richard.case@suse.com>